fix price to double

+++ server/models/models.js
@@ -20,7 +20,7 @@ const BasketDevice = sequelize.define('basket_device', {
 const Device = sequelize.define('device', {
     id: {type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true},
     name: {type: DataTypes.STRING, unique: true, allowNull: false},
-    price: {type: DataTypes.INTEGER, allowNull: false},
+    price: {type: DataTypes.DECIMAL(10, 2), allowNull: false},
     rating: {type: DataTypes.INTEGER, defaultValue: 0},
     img: {type: DataTypes.STRING, allowNull: false},
     img2: {type: DataTypes.STRING, allowNull: true},
